SocialSaints 2/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.6.0
asn1crypto==1.5.1
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.9.1
branca==0.5.0
bs4==0.0.1
certifi==2020.6.20
cffi==1.15.0
channels==3.0.5
chardet==3.0.4
click==8.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==39.0.2
cssselect2==0.5.0
cycler==0.11.0
daphne==3.0.2
dataclasses==0.6
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.18
django-admin-thumbnails==0.2.5
django-ckeditor==5.9.0
django-crispy-forms==1.14.0
django-currencies==0.10.1
django-jazzmin==2.5.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-phonenumber-field==7.0.0
django-social-share==2.3.0
djangorestframework==3.11.0
folium==0.12.1.post1
fonttools==4.37.1
future==0.18.2
geocoder==1.38.1
html5lib==1.1
hyperlink==21.0.0
idna==2.9
image==1.5.33
incremental==22.10.0
Jinja2==3.1.2
jsonfield==2.1.1
kiwisolver==1.4.4
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.0
oauthlib==3.2.0
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
phonenumbers==8.12.57
Pillow==9.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pymongo==3.10.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-gettext==4.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
ratelim==0.1.6
reportlab==3.6.9
requests==2.24.0
requests-oauthlib==1.3.1
seaborn==0.12.0
service-identity==21.1.0
six==1.15.0
social==0.1.dev3
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.0.1
sqlparse==0.3.1
svglib==1.2.1
text-unidecode==1.3
tinycss2==1.1.1
tqdm==4.64.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.25.9
webencodings==0.5.1
xhtml2pdf==0.2.7
zope.interface==5.5.2
